+++ include/table.cuda.h
@@ -7,7 +7,7 @@
 
     typedef table_t cudatable_t;
 
-    cudatable_t* make_cudatable(uint64_t n, uint64_t m);
+    cudatable_t* make_cudatable(int32_t n, int32_t m);
     cudatable_t* copy_cudatable(table_t* cpu_table);
     void free_cudatable(cudatable_t* t);
